+++ src/Oro/Bundle/CMSBundle/DependencyInjection/OroCMSExtension.php
@@ -24,6 +24,7 @@ class OroCMSExtension extends Extension
 
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
         $loader->load('services.yml');
+        $loader->load('services_api.yml');
         $loader->load('form_types.yml');
         $loader->load('block_types.yml');
         $loader->load('controllers.yml');